Breaking changes
* Add imageio as dependency

* Entirely rewritten policy gradient examples

* Update Agent: better notation for configuration string, add save/load of network parameters; all agents handle batched observation even if batch size one,  in choose_action and learn.

* Update CurvePlot: support options with given axes to plot on and xvalues for ticks in horizontal axis; support keyword arguments for legend location, maximum number of horizontal ticks, and formatter of ticks in horizontal axis

* Move Logger from lagom.core.utils to top level

* Update Engine: add independent logging functions for trainng and evaluation

* Update Env: add @property max episode horizon and max episode reward

* Update BaseExperimentMaster: allow number of worker to be None and by default open one Process for each possible configuration

* ExperimentWorker: print configuration in the master, otherwise, strings from several works will mess together

* Update Config: add method to produce a Pandas DataFrame for given list of configuration; add method to partition configuration IDs by given keys, internally use Pandas groupby, this is very useful for plotting and grouping different configurations.

* Update run_experiment: automatically create logging directories and one subdirectory for each configuration named with its ID. It also detects existing logging folder and ask use whether to erase it or back it up

* Update Logger: entirely rewritten, because previously we allow hierarchical structure, e.g. list of dict of list of dict of ndarray, it can be super slow to save it to disk. So this new version we only allow top level dictionary. And a new dump function to automatically dump logged information to the screen with the option of index and indentation

* Update TrajectoryRunner: Handle everything in batches

* Add SegmentRunner: useful for batched algorithm e.g. A2C


Former-commit-id: ffe87a3e0b8d696b29fde1be57f8e68782acdcf1 [formerly 177a9cf83ba69020a5508b4412a0a7f6beb0a91c]
Former-commit-id: befa4e37ee4931c400ba27ced74563ac37075607

+++ lagom/runner/trajectory.py
@@ -140,7 +140,7 @@ class Trajectory(object):
         # Calculate TD error
         all_TD = all_r + self.gamma*all_V_s_next - all_V_s
         
-        return all_TD.tolist()
+        return all_TD.astype(np.float32).tolist()
     
     @property
     def all_gae(self, gae_lambda):
